+++ src/tests/unittests/configuration_helper/adapters/test_d5a_instrument_adapter.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 
 from qilib.configuration_helper import InstrumentAdapterFactory, SerialPortResolver
 from qilib.configuration_helper.adapters import D5aInstrumentAdapter
-from qilib.configuration_helper.adapters.common_config_check_instrument_adapter import ConfigurationError
+from qilib.configuration_helper.adapters.common_config_instrument_adapter import ConfigurationError
 from qilib.configuration_helper.adapters.d5a_instrument_adapter import SpanValueError
