+++ src/tests/unittests/configuration_helper/test_insturment_configuration.py
@@ -1,6 +1,5 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from typing import Any
 
 from qilib.configuration_helper import InstrumentConfiguration
 from qilib.configuration_helper.exceptions import DuplicateTagError
@@ -33,9 +32,6 @@ class TestInstrumentConfiguration(unittest.TestCase):
                 pass
 
             def _filter_parameters(self, parameters: PythonJsonStructure) -> PythonJsonStructure:
-                pass
-
-            def _compare_config_values(self, config_value: Any, device_value: Any, parameter: str) -> bool:
                 pass
